Update build scripts

+++ build.sh
@@ -1,17 +1,69 @@
 #!/usr/bin/env sh
 
 family=SourceSerifPro
-romanWeights='ExtraLight Light Regular Semibold Bold Black'
+roman_weights='ExtraLight Light Regular Semibold Bold Black'
+italic_weights='ExtraLightIt LightIt It SemiboldIt BoldIt BlackIt'
+fonttools_dir=~/code/fonttoolse
+comp_script=$fonttools_dir/Snippets/otf2ttf.py
+
+script_found=True
+if [[ ! -f $comp_script ]]; then
+	echo "ERROR: Cannot find fontTools directory snippet at"
+	echo $fonttools_dir
+	echo "Please get fontTools at https://github.com/fonttools/fonttools"
+	echo "and edit the 'fonttools_dir' path in this build script."
+	script_found=false
+fi
+
+if ! $script_found; then
+	exit 1
+fi
 
 # clean existing build artifacts
 rm -rf target/
-otfDir="target/OTF"
-ttfDir="target/TTF"
-mkdir -p $otfDir $ttfDir
+otf_dir="target/OTF"
+ttf_dir="target/TTF"
+mkdir -p $otf_dir $ttf_dir
 
-for w in $romanWeights
+for w in $roman_weights
 do
-  font_path=Roman/Instances/$w/font
-  makeotf -f $font_path.ufo -r -o $otfDir/$family-$w.otf
-  makeotf -f $font_path.ttf -r -o $ttfDir/$family-$w.ttf -ff $font_path.ufo/features.fea
+  font_dir=Roman/Instances/$w
+  font_ufo=$font_dir/font.ufo
+  ps_name=$family-$w
+
+  echo "Building OTF ..."
+  makeotf -f $font_ufo -r -o $font_dir/$ps_name.otf
+  echo "Building TTF ..."
+  python3 $comp_script $font_dir/$ps_name.otf -o $font_dir/$ps_name.ttf
+  echo "Componentizing TTF ..."
+  ttfcomponentizer $font_dir/$ps_name.ttf
+
+  # move font files to target directory
+  mv $font_dir/$ps_name.otf $otf_dir/
+  mv $font_dir/$ps_name.ttf $ttf_dir/
+  echo "Done with $ps_name"
+  echo ""
+  echo ""
+done
+
+
+for w in $italic_weights
+do
+  font_dir=Italic/Instances/$w
+  font_ufo=$font_dir/font.ufo
+  ps_name=$family-$w
+
+  echo "Building OTF ..."
+  makeotf -f $font_ufo -r -o $font_dir/$ps_name.otf
+  echo "Building TTF ..."
+  python3 $comp_script $font_dir/$ps_name.otf -o $font_dir/$ps_name.ttf
+  echo "Componentizing TTF ..."
+  ttfcomponentizer $font_dir/$ps_name.ttf
+
+  # move font files to target directory
+  mv $font_dir/$ps_name.otf $otf_dir/
+  mv $font_dir/$ps_name.ttf $ttf_dir/
+  echo "Done with $ps_name"
+  echo ""
+  echo ""
 done
